Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -5,7 +5,7 @@ expenses = int(input("Enter your total monthly expenses: "))
 
 monthly_saving = income - expenses
 
-projected_saving = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
+projected_saving = int((monthly_saving * 12 + (monthly_saving * 12 * 0.05)))
 
 print(f"Your monthly savings are ${monthly_saving}.")
 print(f"Projected savings after one year, with interest, is: ${projected_saving}.")
